Modify encoder's entropy coding byte-push mechanism
Changed the byte-push mechanism in libaom entropy
encoder. In the parent version, the register is
flushed when at least 1 byte is available. In the
CL, we make the register 64-bit and flush it when
at least 4 bytes are accumulated in it.

An intermediate 16-bit buffer 'precarry_buf' is
removed and carry-propagation mechanism is changed.

No changes to the decoder logic.

<Update speed numbers>

+++ aom_dsp/entcode.h
@@ -22,7 +22,7 @@
 
 /*OPT: od_ec_window must be at least 32 bits, but if you have fast arithmetic
    on a larger type, you can speed up the decoder by using it here.*/
-typedef uint32_t od_ec_window;
+typedef uint64_t od_ec_window;
 
 /*The size in bits of od_ec_window.*/
 #define OD_EC_WINDOW_SIZE ((int)sizeof(od_ec_window) * CHAR_BIT)
